Add enhancements to HTML, CSS and JS Auth files
Implement show/hide of auth and game content.
Only sign in and sign up display on initial load.
Everything else displays on successful log in.
On successful log in, sign in and sign up hide.
Change password lives in a modal at bottom of screen.
Sign out also on bottom of screen.